refactor: extract background upload discovery
