+++ BackgroundUploadExtension/BackgroundUploadExtension+Destination.swift
@@ -46,40 +46,22 @@ extension BackgroundUploadExtension {
         var request = URLRequest(url: url)
         request.httpMethod = "PUT"
 
-        request.setValue(
-            nkSession.userAgent,
-            forHTTPHeaderField: "User-Agent"
-        )
+        request.setValue(nkSession.userAgent, forHTTPHeaderField: "User-Agent")
 
-        request.setValue(
-            "Basic \(loginData.base64EncodedString())",
-            forHTTPHeaderField: "Authorization"
-        )
+        request.setValue("Basic \(loginData.base64EncodedString())", forHTTPHeaderField: "Authorization")
 
-        request.setValue(
-            "application/octet-stream",
-            forHTTPHeaderField: "Content-Type"
-        )
+        request.setValue("application/octet-stream", forHTTPHeaderField: "Content-Type")
 
-        request.setValue(
-            "1",
-            forHTTPHeaderField: "X-NC-WebDAV-Auto-Mkcol"
-        )
+        request.setValue("1", forHTTPHeaderField: "X-NC-WebDAV-Auto-Mkcol")
 
         if let creationDate = asset.creationDate,
            creationDate.timeIntervalSince1970 > 0 {
-            request.setValue(
-                "\(creationDate.timeIntervalSince1970)",
-                forHTTPHeaderField: "X-OC-CTime"
-            )
+            request.setValue("\(creationDate.timeIntervalSince1970)", forHTTPHeaderField: "X-OC-CTime")
         }
 
         if let modificationDate = asset.modificationDate,
            modificationDate.timeIntervalSince1970 > 0 {
-            request.setValue(
-                "\(modificationDate.timeIntervalSince1970)",
-                forHTTPHeaderField: "X-OC-MTime"
-            )
+            request.setValue("\(modificationDate.timeIntervalSince1970)", forHTTPHeaderField: "X-OC-MTime")
         }
 
         nkLog(
